+++ crates/symbolicator/src/services/minidump.rs
@@ -96,10 +96,9 @@
 use std::convert::TryFrom;
 use std::fmt;
 
-use symbolic::minidump::processor::FrameTrust;
 use thiserror::Error;
 
-use crate::types;
+use crate::types::{self, FrameTrust};
 use crate::utils::hex;
 
 const MINIDUMP_EXTENSION_TYPE: u32 = u32::from_be_bytes([b'S', b'y', 0, 1]);
@@ -120,7 +119,7 @@ pub fn parse_stacktraces_from_minidump(
     let stacktraces = parsed
         .threads()
         .map(types::RawStacktrace::try_from)
-        .collect::<Result<Vec<_>, _>>()?;
+        .collect::<Result<_, _>>()?;
 
     Ok(Some(stacktraces))
 }
@@ -156,7 +155,7 @@ impl TryFrom<format::Frame<'_>> for types::RawFrame {
         Ok(types::RawFrame {
             instruction_addr: hex::HexValue(frame.instruction_addr()),
             function: Some(String::from_utf8_lossy(symbol).into_owned()),
-            trust: FrameTrust::Prewalked,
+            trust: FrameTrust::PreWalked,
             ..Default::default()
         })
     }